refactor (moved run_verse out) UI not updated yet with new settings
